pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bdd_vision"]

[project]
name = "bdd-vision"
version = "0.1.0"
description = "AI-powered BDD testing agent using vision-language models"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "pillow>=10.0",
    "mss>=9.0",
    "pyautogui>=0.9.54",
    "reportlab>=4.0",
    "python-dotenv>=1.0",
    "google-generativeai>=0.8",
    "anthropic>=0.40",
    "websockets>=13.0",
    "rich>=13.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
bdd-vision = "bdd_vision.cli.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
